# OPENAPI.yaml
openapi: 3.0.3
info:
  title: Customer Voice Dashboard API
  version: 1.0.0
  description: >
    REST API powering the Customer Voice Dashboard. Provides review ingestion,
    on-demand analysis, insights aggregation, competitor management, and digest generation.
servers:
  - url: https://api.customer-voice.example.com
    description: Production
  - url: http://localhost:5000
    description: Local development
tags:
  - name: Ingest
  - name: Analysis
  - name: Insights
  - name: Competitors
  - name: Digest
paths:
  /ingest:
    post:
      tags: [Ingest]
      summary: Batch ingest customer reviews
      operationId: postIngest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewIngestRequest'
      responses:
        '202':
          description: Reviews accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewIngestResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /analyze:
    post:
      tags: [Analysis]
      summary: Analyze free-form text for sentiment and topics
      operationId: postAnalyze
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '200':
          description: Analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /insights:
    get:
      tags: [Insights]
      summary: Retrieve aggregated insights across reviews
      operationId: getInsights
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/StartDateParam'
        - $ref: '#/components/parameters/EndDateParam'
        - $ref: '#/components/parameters/SourceIdParam'
        - $ref: '#/components/parameters/SentimentFilterParam'
      responses:
        '200':
          description: Aggregated insight payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsightsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /competitors:
    get:
      tags: [Competitors]
      summary: List tracked competitors
      operationId: listCompetitors
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of competitors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitorListResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      tags: [Competitors]
      summary: Create a competitor record
      operationId: createCompetitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitorCreate'
      responses:
        '201':
          description: Competitor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competitor'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /competitors/{competitorId}:
    parameters:
      - $ref: '#/components/parameters/CompetitorIdPath'
    get:
      tags: [Competitors]
      summary: Retrieve competitor details
      operationId: getCompetitor
      responses:
        '200':
          description: Competitor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competitor'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    patch:
      tags: [Competitors]
      summary: Update competitor metadata
      operationId: updateCompetitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitorUpdate'
      responses:
        '200':
          description: Updated competitor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competitor'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      tags: [Competitors]
      summary: Delete a competitor
      operationId: deleteCompetitor
      responses:
        '204':
          description: Competitor removed
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /competitors/{competitorId}/comparison:
    parameters:
      - $ref: '#/components/parameters/CompetitorIdPath'
      - $ref: '#/components/parameters/StartDateParam'
      - $ref: '#/components/parameters/EndDateParam'
    get:
      tags: [Competitors]
      summary: Compare competitor metrics against our product
      operationId: getCompetitorComparison
      responses:
        '200':
          description: Sentiment and topic comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitorComparisonResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /digest/run:
    post:
      tags: [Digest]
      summary: Generate the latest digest
      operationId: runDigest
      security:
        - BearerToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DigestRequest'
      responses:
        '200':
          description: Digest generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Static bearer token stored in TOKEN_DIGEST_RUN
  parameters:
    PageParam:
      name: page
      in: query
      description: 1-based page index
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    StartDateParam:
      name: start_date
      in: query
      description: Filter results on or after this ISO-8601 date
      schema:
        type: string
        format: date
    EndDateParam:
      name: end_date
      in: query
      description: Filter results on or before this ISO-8601 date
      schema:
        type: string
        format: date
    SourceIdParam:
      name: source_id
      in: query
      description: Restrict insights to a specific source UUID
      schema:
        type: string
        format: uuid
    SentimentFilterParam:
      name: sentiment
      in: query
      description: Filter by sentiment label
      schema:
        $ref: '#/components/schemas/SentimentLabel'
    CompetitorIdPath:
      name: competitorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests; retry after the supplied delay
      headers:
        Retry-After:
          description: Seconds until another request is allowed
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
  schemas:
    SentimentLabel:
      type: string
      enum: [Positive, Neutral, Negative]
    ReviewIngestRequest:
      type: object
      required: [source_id, reviews]
      properties:
        source_id:
          type: string
          format: uuid
        overwrite_source_metadata:
          type: boolean
          default: false
          description: Apply provided source metadata to update the source record.
        source_metadata:
          $ref: '#/components/schemas/SourceMetadata'
        reviews:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ReviewIngestItem'
    SourceMetadata:
      type: object
      properties:
        external_id:
          type: string
        name:
          type: string
        platform:
          type: string
          description: e.g. app_store, play_store, zendesk
        url:
          type: string
          format: uri
    ReviewIngestItem:
      type: object
      required:
        - source_review_id
        - body
        - rating
        - published_at
      properties:
        source_review_id:
          type: string
          description: Unique ID supplied by the review source.
        title:
          type: string
        body:
          type: string
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        author_name:
          type: string
        language:
          type: string
          description: ISO 639-1 code
        location:
          type: string
        published_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    ReviewIngestResponse:
      type: object
      required: [ingested_count, duplicate_count, review_ids]
      properties:
        ingested_count:
          type: integer
          minimum: 0
        duplicate_count:
          type: integer
          minimum: 0
        review_ids:
          type: array
          items:
            type: string
            format: uuid
        message:
          type: string
    AnalyzeRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
        language:
          type: string
          description: Optional ISO 639-1 language hint
    AnalyzeResponse:
      type: object
      required: [sentiment, topics]
      properties:
        sentiment:
          $ref: '#/components/schemas/Sentiment'
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicScore'
    Sentiment:
      type: object
      required: [label, score]
      properties:
        label:
          $ref: '#/components/schemas/SentimentLabel'
        score:
          type: number
          format: float
          minimum: -1
          maximum: 1
    TopicScore:
      type: object
      required: [topic_label, topic_confidence]
      properties:
        topic_label:
          type: string
        topic_confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
    InsightsResponse:
      type: object
      required:
        - pagination
        - sentiment_trend
        - topic_distribution
        - source_breakdown
        - recent_reviews
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        sentiment_trend:
          type: array
          items:
            $ref: '#/components/schemas/SentimentTrendPoint'
        topic_distribution:
          type: array
          items:
            $ref: '#/components/schemas/TopicDistributionItem'
        source_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/SourceBreakdownItem'
        recent_reviews:
          type: array
          items:
            $ref: '#/components/schemas/RecentReview'
    SentimentTrendPoint:
      type: object
      required: [date, positive, neutral, negative]
      properties:
        date:
          type: string
          format: date
        positive:
          type: integer
          minimum: 0
        neutral:
          type: integer
          minimum: 0
        negative:
          type: integer
          minimum: 0
        average_score:
          type: number
          format: float
          minimum: -1
          maximum: 1
    TopicDistributionItem:
      type: object
      required: [topic_label, review_count, average_confidence]
      properties:
        topic_label:
          type: string
        review_count:
          type: integer
          minimum: 0
        average_confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
    SourceBreakdownItem:
      type: object
      required: [source_id, source_name, review_count, average_sentiment_score]
      properties:
        source_id:
          type: string
          format: uuid
        source_name:
          type: string
        review_count:
          type: integer
          minimum: 0
        average_sentiment_score:
          type: number
          format: float
          minimum: -1
          maximum: 1
    RecentReview:
      type: object
      required:
        - review_id
        - source_id
        - source_review_id
        - title
        - sentiment
        - published_at
      properties:
        review_id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        source_review_id:
          type: string
        title:
          type: string
        body:
          type: string
        sentiment:
          $ref: '#/components/schemas/Sentiment'
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicScore'
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        published_at:
          type: string
          format: date-time
    Pagination:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
    CompetitorListResponse:
      type: object
      required: [pagination, items]
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Competitor'
    Competitor:
      type: object
      required:
        - competitor_id
        - name
        - created_at
        - updated_at
      properties:
        competitor_id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CompetitorCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    CompetitorUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    CompetitorComparisonResponse:
      type: object
      required:
        - competitor
        - self_sentiment
        - competitor_sentiment
        - top_topics
      properties:
        competitor:
          $ref: '#/components/schemas/Competitor'
        self_sentiment:
          $ref: '#/components/schemas/SentimentSummary'
        competitor_sentiment:
          $ref: '#/components/schemas/SentimentSummary'
        top_topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicComparison'
    SentimentSummary:
      type: object
      required: [positive, neutral, negative, average_score, review_count]
      properties:
        positive:
          type: integer
          minimum: 0
        neutral:
          type: integer
          minimum: 0
        negative:
          type: integer
          minimum: 0
        average_score:
          type: number
          format: float
          minimum: -1
          maximum: 1
        review_count:
          type: integer
          minimum: 0
    TopicComparison:
      type: object
      required: [topic_label, self_share, competitor_share, delta]
      properties:
        topic_label:
          type: string
        self_share:
          type: number
          format: float
        competitor_share:
          type: number
          format: float
        delta:
          type: number
          format: float
    DigestRequest:
      type: object
      properties:
        timeframe_start:
          type: string
          format: date-time
        timeframe_end:
          type: string
          format: date-time
        include_competitors:
          type: boolean
          default: true
    DigestResponse:
      type: object
      required:
        - digest_id
        - timeframe_start
        - timeframe_end
        - generated_at
        - highlights
        - key_metrics
      properties:
        digest_id:
          type: string
          format: uuid
        timeframe_start:
          type: string
          format: date-time
        timeframe_end:
          type: string
          format: date-time
        generated_at:
          type: string
          format: date-time
        highlights:
          type: array
          items:
            type: string
        key_metrics:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
        sentiment_snapshot:
          $ref: '#/components/schemas/SentimentSummary'
        topic_spotlight:
          type: array
          items:
            $ref: '#/components/schemas/TopicSpotlight'
        competitor_summary:
          type: array
          items:
            $ref: '#/components/schemas/CompetitorDigestItem'
    TopicSpotlight:
      type: object
      required: [topic_label, change_vs_previous, sample_quotes]
      properties:
        topic_label:
          type: string
        change_vs_previous:
          type: number
          format: float
        sample_quotes:
          type: array
          items:
            type: string
    CompetitorDigestItem:
      type: object
      required: [competitor_id, name, sentiment_delta]
      properties:
        competitor_id:
          type: string
          format: uuid
        name:
          type: string
        sentiment_delta:
          type: number
          format: float
        highlight:
          type: string
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
    RateLimitError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required: [retry_after_seconds, guidance]
          properties:
            retry_after_seconds:
              type: integer
              minimum: 1
            guidance:
              type: string